Added Bonded Support Prototype
Added a bonded support prototype for the Pixel 2 XL support. 
It should be easy to place on the center dash. The original support which had to be drilled can then be fixed onto the support (or screwed probably). 
There is a channel underneath for the usb c cable and some marks for double sided tape. 
This is a prototype and still needs to be tested in the real world. The fitment on the dash should work as well as the fitment on the Pixel 2 support.